# wwwdir/venom/shared/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Api
  description: VENOM CODES API specification — WHMCS-backed client portal
  version: 0.1.0
servers:
  - url: /
paths:
...
  /api/client/profile:
    get:
      tags: [Client]
      summary: Get client profile
      operationId: getClientProfile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
    put:
      tags: [Client]
      summary: Update client profile
      operationId: updateClientProfile
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
  /api/client/contacts:
    get:
      tags: [Client]
      summary: List contacts
      operationId: getClientContacts
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      tags: [Client]
      summary: Add contact
      operationId: addClientContact
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/client/contacts/{contactId}:
    put:
      tags: [Client]
      summary: Update contact
      operationId: updateClientContact
      security: [{ bearerAuth: [] }]
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
    delete:
      tags: [Client]
      summary: Delete contact
      operationId: deleteClientContact
      security: [{ bearerAuth: [] }]
      parameters:
        - name: contactId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/client/change-password:
    post:
      tags: [Client]
      summary: Change password
      operationId: changePassword
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/client/email-preferences:
    get:
      tags: [Client]
      summary: Get email preferences
      operationId: getEmailPreferences
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailPreferences"
    put:
      tags: [Client]
      summary: Update email preferences
      operationId: updateEmailPreferences
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailPreferences"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/client/credits:
    get:
      tags: [Client]
      summary: Client credit balance
      operationId: getClientCredits
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
  /api/client/dashboard:
    get:
      tags: [Client]
      summary: Dashboard summary
      operationId: getDashboardSummary
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
...
  /api/invoices:
    get:
      tags: [Billing]
      summary: List invoices
      operationId: getInvoices
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
...
  /api/tickets:
    get:
      tags: [Support]
      summary: List tickets
      operationId: getTickets
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: deptid
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketListResponse"
    post:
      tags: [Support]
      summary: Open ticket
      operationId: createTicket
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
  /api/tickets/counts:
    get:
      tags: [Support]
      summary: WHMCS GetTicketCounts
      operationId: getTicketCounts
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnyJson"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AnyJson:
      type: object
      additionalProperties: true
...
    ClientProfile:
      type: object
      required: [id, firstname, lastname, email, status]
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        status:
          type: string
        companyname:
          type: string
          nullable: true
        phonenumber:
          type: string
        address1:
          type: string
        address2:
          type: string
          nullable: true
        city:
          type: string
        state:
          type: string
        postcode:
          type: string
        country:
          type: string
        currency:
          type: number
        currencyCode:
          type: string
        credit:
          type: string
        createdAt:
          type: string
    UpdateClientRequest:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
    Contact:
      type: object
      required: [id, firstname, lastname, email]
      properties:
        id:
          type: string
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        phonenumber:
          type: string
        subaccount:
          type: boolean
    ContactInput:
      type: object
      required: [firstname, lastname, email]
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
        phonenumber:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
    EmailPreferences:
      type: object
      required:
        - generalEmails
        - invoiceEmails
        - productEmails
        - domainEmails
        - supportEmails
        - marketingEmails
      properties:
        generalEmails:
          type: boolean
        invoiceEmails:
          type: boolean
        productEmails:
          type: boolean
        domainEmails:
          type: boolean
        supportEmails:
          type: boolean
        marketingEmails:
          type: boolean
    DashboardSummary:
      type: object
    CreateTicketRequest:
      type: object
      required: [departmentId, subject, message]
      properties:
        departmentId:
          type: string
        subject:
          type: string
        message:
          type: string
        priority:
          type: string
    InvoiceListItem:
      type: object
      required: [id, status, total]
      properties:
        id:
          type: string
        status:
          type: string
        date:
          type: string
        dueDate:
          type: string
        total:
          type: string
        subtotal:
          type: string
        taxAmount:
          type: string
        credit:
          type: string
        paymentMethod:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    InvoiceListResponse:
      type: object
      required: [invoices, total]
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceListItem"
        total:
          type: number
...
    TicketSummary:
      type: object
      required: [id, subject, status, priority]
      properties:
        id:
          type: string
        subject:
          type: string
        status:
          type: string
        department:
          type: string
        priority:
          type: string
        lastUpdated:
          type: string
    TicketListResponse:
      type: object
      required: [tickets]
      properties:
        tickets:
          type: array
          items:
            $ref: "#/components/schemas/TicketSummary"
...